pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dashscope>=1.24.0",
    "faiss-cpu>=1.11.0.post1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-deepseek>=0.1.4",
    "langchain-mcp-adapters>=0.1.9",
    "mcp>=1.12.2",
    "openai>=1.97.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]

[tool.uv.workspace]
members = [
    "mcp-client",
]
